Refactored magda-gateway etc.
