+++ magda-registry-api/src/test/scala/au/csiro/data61/magda/registry/RecordsServiceSpec.scala
@@ -34,14 +34,14 @@ class RecordsServiceSpec extends ApiSpec {
   def readOnlyTests(role: Role) {
     describe("GET") {
       it("starts with no records defined") { param =>
-        Get("/v0/records") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        Get("/v0/records") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
           responseAs[RecordsPage[Record]].records shouldBe empty
         }
       }
 
       it("returns 404 if the given ID does not exist") { param =>
-        Get("/v0/records/foo") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        Get("/v0/records/foo") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.NotFound
           responseAs[BadRequest].message should include("exist")
         }
@@ -49,16 +49,16 @@ class RecordsServiceSpec extends ApiSpec {
 
       it("returns 404 if the given ID does not have a required aspect") { param =>
         val aspectDefinition = AspectDefinition("test", "test", None)
-        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val record = Record("foo", "foo", Map())
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
-        Get("/v0/records/foo?aspect=test") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        Get("/v0/records/foo?aspect=test") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.NotFound
           responseAs[BadRequest].message should include("exist")
         }
@@ -69,16 +69,20 @@ class RecordsServiceSpec extends ApiSpec {
           insertAspectDefs(param)
 
           val recordWithAspects = Record("id", "name", Map("test1" -> JsObject(), "test2" -> JsObject()))
-          param.asAdmin(Post("/v0/records", recordWithAspects)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", recordWithAspects)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          Get("/v0/records/summary/id") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records/summary/id") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val recordSummary = responseAs[RecordSummary]
             recordSummary.id shouldEqual "id"
             recordSummary.name shouldEqual "name"
             recordSummary.aspects shouldEqual List("test2", "test1")
+          }
+
+          Get("/v0/records/summary/id") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.NotFound
           }
         }
 
@@ -86,21 +90,21 @@ class RecordsServiceSpec extends ApiSpec {
           insertAspectDefs(param)
 
           val recordWithAspects1 = Record("id1", "name1", Map("test1" -> JsObject(), "test2" -> JsObject()))
-          param.asAdmin(Post("/v0/records", recordWithAspects1)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", recordWithAspects1)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val recordWithAspects2 = Record("id2", "name2", Map("test2" -> JsObject(), "test3" -> JsObject()))
-          param.asAdmin(Post("/v0/records", recordWithAspects2)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", recordWithAspects2)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val recordWithoutAspects = Record("id3", "name3", Map())
-          param.asAdmin(Post("/v0/records", recordWithoutAspects)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", recordWithoutAspects)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          Get("/v0/records/summary") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records/summary") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val recordsSummary = responseAs[RecordsPage[RecordSummary]]
             recordsSummary.records.length shouldEqual 3
@@ -117,19 +121,25 @@ class RecordsServiceSpec extends ApiSpec {
             recordsSummary.records(2).name shouldEqual "name3"
             recordsSummary.records(2).aspects shouldEqual List()
           }
+
+          Get("/v0/records/summary") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            val recordsSummary = responseAs[RecordsPage[RecordSummary]]
+            recordsSummary.records.length shouldEqual 0
+          }
         }
 
         def insertAspectDefs(param: FixtureParam) {
           val aspectDefinition1 = AspectDefinition("test1", "test1", None)
-          param.asAdmin(Post("/v0/aspects", aspectDefinition1)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", aspectDefinition1)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
           val aspectDefinition2 = AspectDefinition("test2", "test2", None)
-          param.asAdmin(Post("/v0/aspects", aspectDefinition2)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", aspectDefinition2)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
           val aspectDefinition3 = AspectDefinition("test3", "test3", None)
-          param.asAdmin(Post("/v0/aspects", aspectDefinition3)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", aspectDefinition3)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
         }
@@ -139,15 +149,21 @@ class RecordsServiceSpec extends ApiSpec {
         it("returns the right count when no parameters are given") { param =>
           for (i <- 1 to 5) {
             val recordWithoutAspects = Record("id" + i, "name" + i, Map())
-            param.asAdmin(Post("/v0/records", recordWithoutAspects)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+            param.asAdmin(Post("/v0/records", recordWithoutAspects)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
               status shouldEqual StatusCodes.OK
             }
           }
 
-          Get("/v0/records/count") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records/count") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val countResponse = responseAs[CountResponse]
             countResponse.count shouldBe 5
+          }
+
+          Get("/v0/records/count") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            val countResponse = responseAs[CountResponse]
+            countResponse.count shouldBe 0
           }
         }
       }
@@ -155,272 +171,366 @@ class RecordsServiceSpec extends ApiSpec {
       describe("aspects") {
         it("includes optionalAspect if it exists") { param =>
           val aspectDefinition = AspectDefinition("test", "test", None)
-          param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          val recordWithAspect = Record("with", "with", Map("test" -> JsObject("foo" -> JsString("bar"))))
-          param.asAdmin(Post("/v0/records", recordWithAspect)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          val recordWithAspect = Record("with", "with", Map("test" -> JsObject("foo" -> JsString("bar"))), tenantId = Some(TENANT_1))
+          param.asAdmin(Post("/v0/records", recordWithAspect)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          val recordWithoutAspect = Record("without", "without", Map())
-          param.asAdmin(Post("/v0/records", recordWithoutAspect)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          val recordWithoutAspect = Record("without", "without", Map(), tenantId = Some(TENANT_1))
+          param.asAdmin(Post("/v0/records", recordWithoutAspect)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          Get("/v0/records?optionalAspect=test") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records?optionalAspect=test") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val page = responseAs[RecordsPage[Record]]
             page.records.length shouldBe 2
-            page.records.head shouldBe recordWithAspect.copy(tenantId = Some(SINGLE_TENANT_ID))
-            page.records(1) shouldBe recordWithoutAspect.copy(tenantId = Some(SINGLE_TENANT_ID))
+            page.records.head shouldBe recordWithAspect
+            page.records(1) shouldBe recordWithoutAspect
+          }
+
+          Get("/v0/records?optionalAspect=test") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            val page = responseAs[RecordsPage[Record]]
+            page.records.length shouldBe 0
           }
         }
 
         it("requires presence of aspect") { param =>
           val aspectDefinition = AspectDefinition("test", "test", None)
-          param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          val recordWithAspect = Record("with", "with", Map("test" -> JsObject("foo" -> JsString("bar"))))
-          param.asAdmin(Post("/v0/records", recordWithAspect)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          val recordWithAspect = Record("with", "with", Map("test" -> JsObject("foo" -> JsString("bar"))), tenantId = Some(TENANT_1))
+          param.asAdmin(Post("/v0/records", recordWithAspect)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          val recordWithoutAspect = Record("without", "without", Map())
-          param.asAdmin(Post("/v0/records", recordWithoutAspect)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          val recordWithoutAspect = Record("without", "without", Map(), tenantId = Some(TENANT_1))
+          param.asAdmin(Post("/v0/records", recordWithoutAspect)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          Get("/v0/records?aspect=test") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records?aspect=test") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val page = responseAs[RecordsPage[Record]]
             page.records.length shouldBe 1
-            page.records.head shouldBe recordWithAspect.copy(tenantId = Some(SINGLE_TENANT_ID))
+            page.records.head shouldBe recordWithAspect
           }
 
-          Get("/v0/records/count?aspect=test") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records/count?aspect=test") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val countResponse = responseAs[CountResponse]
             countResponse.count shouldBe 1
+          }
+
+          Get("/v0/records?aspect=test") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            val page = responseAs[RecordsPage[Record]]
+            page.records.length shouldBe 0
+          }
+
+          Get("/v0/records/count?aspect=test") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            val countResponse = responseAs[CountResponse]
+            countResponse.count shouldBe 0
           }
         }
 
         it("requires any specified aspects to be present") { param =>
           val fooAspect = AspectDefinition("foo", "foo", None)
-          param.asAdmin(Post("/v0/aspects", fooAspect)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", fooAspect)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val barAspect = AspectDefinition("bar", "bar", None)
-          param.asAdmin(Post("/v0/aspects", barAspect)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", barAspect)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val withFoo = Record("withFoo", "with foo", Map("foo" -> JsObject("test" -> JsString("test"))))
-          param.asAdmin(Post("/v0/records", withFoo)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", withFoo)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val withBar = Record("withBar", "with bar", Map("bar" -> JsObject("test" -> JsString("test"))))
-          param.asAdmin(Post("/v0/records", withBar)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", withBar)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          val withFooAndBar = Record("withFooAndBar", "with foo and bar", Map("foo" -> JsObject(), "bar" -> JsObject("test" -> JsString("test"))))
-          param.asAdmin(Post("/v0/records", withFooAndBar)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          val withFooAndBar = Record("withFooAndBar", "with foo and bar", Map("foo" -> JsObject(), "bar" -> JsObject("test" -> JsString("test"))), tenantId = Some(TENANT_1))
+          param.asAdmin(Post("/v0/records", withFooAndBar)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          Get("/v0/records?aspect=foo&aspect=bar") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records?aspect=foo&aspect=bar") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val page = responseAs[RecordsPage[Record]]
             page.records.length shouldBe 1
-            page.records.head shouldBe withFooAndBar.copy(tenantId = Some(SINGLE_TENANT_ID))
+            page.records.head shouldBe withFooAndBar
           }
 
-          Get("/v0/records/count?aspect=foo&aspect=bar") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records/count?aspect=foo&aspect=bar") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val countResponse = responseAs[CountResponse]
             countResponse.count shouldBe 1
+          }
+
+          Get("/v0/records?aspect=foo&aspect=bar") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            val page = responseAs[RecordsPage[Record]]
+            page.records.length shouldBe 0
+          }
+
+          Get("/v0/records/count?aspect=foo&aspect=bar") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            val countResponse = responseAs[CountResponse]
+            countResponse.count shouldBe 0
           }
         }
 
         it("optionalAspects are optional") { param =>
           val fooAspect = AspectDefinition("foo", "foo", None)
-          param.asAdmin(Post("/v0/aspects", fooAspect)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", fooAspect)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val barAspect = AspectDefinition("bar", "bar", None)
-          param.asAdmin(Post("/v0/aspects", barAspect)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", barAspect)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          val withFoo = Record("withFoo", "with foo", Map("foo" -> JsObject("test" -> JsString("test"))))
-          param.asAdmin(Post("/v0/records", withFoo)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          val withFoo = Record("withFoo", "with foo", Map("foo" -> JsObject("test" -> JsString("test"))), tenantId = Some(TENANT_1))
+          param.asAdmin(Post("/v0/records", withFoo)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          val withBar = Record("withBar", "with bar", Map("bar" -> JsObject("test" -> JsString("test"))))
-          param.asAdmin(Post("/v0/records", withBar)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          val withBar = Record("withBar", "with bar", Map("bar" -> JsObject("test" -> JsString("test"))), tenantId = Some(TENANT_1))
+          param.asAdmin(Post("/v0/records", withBar)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          val withFooAndBar = Record("withFooAndBar", "with foo and bar", Map("foo" -> JsObject(), "bar" -> JsObject("test" -> JsString("test"))))
-          param.asAdmin(Post("/v0/records", withFooAndBar)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          val withFooAndBar = Record("withFooAndBar", "with foo and bar", Map("foo" -> JsObject(), "bar" -> JsObject("test" -> JsString("test"))), tenantId = Some(TENANT_1))
+          param.asAdmin(Post("/v0/records", withFooAndBar)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          val withNone = Record("withNone", "with none", Map())
-          param.asAdmin(Post("/v0/records", withNone)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          val withNone = Record("withNone", "with none", Map(), tenantId = Some(TENANT_1))
+          param.asAdmin(Post("/v0/records", withNone)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          Get("/v0/records?optionalAspect=foo&optionalAspect=bar") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records?optionalAspect=foo&optionalAspect=bar") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val page = responseAs[RecordsPage[Record]]
             page.records.length shouldBe 4
-            page.records.head shouldBe withFoo.copy(tenantId = Some(SINGLE_TENANT_ID))
-            page.records(1) shouldBe withBar.copy(tenantId = Some(SINGLE_TENANT_ID))
-            page.records(2) shouldBe withFooAndBar.copy(tenantId = Some(SINGLE_TENANT_ID))
-            page.records(3) shouldBe withNone.copy(tenantId = Some(SINGLE_TENANT_ID))
+            page.records.head shouldBe withFoo
+            page.records(1) shouldBe withBar
+            page.records(2) shouldBe withFooAndBar
+            page.records(3) shouldBe withNone
+          }
+
+          Get("/v0/records?optionalAspect=foo&optionalAspect=bar") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            val page = responseAs[RecordsPage[Record]]
+            page.records.length shouldBe 0
           }
         }
 
         it("supports a mix of aspects and optionalAspects") { param =>
           val fooAspect = AspectDefinition("foo", "foo", None)
-          param.asAdmin(Post("/v0/aspects", fooAspect)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", fooAspect)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val barAspect = AspectDefinition("bar", "bar", None)
-          param.asAdmin(Post("/v0/aspects", barAspect)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", barAspect)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val bazAspect = AspectDefinition("baz", "baz", None)
-          param.asAdmin(Post("/v0/aspects", bazAspect)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", bazAspect)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          val withFoo = Record("withFoo", "with foo", Map("foo" -> JsObject("test" -> JsString("test"))))
-          param.asAdmin(Post("/v0/records", withFoo)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          val withFoo = Record("withFoo", "with foo", Map("foo" -> JsObject("test" -> JsString("test"))), tenantId = Some(TENANT_1))
+          param.asAdmin(Post("/v0/records", withFoo)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val withBar = Record("withBar", "with bar", Map("bar" -> JsObject("test" -> JsString("test"))))
-          param.asAdmin(Post("/v0/records", withBar)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", withBar)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          val withFooAndBarAndBaz = Record("withFooAndBarAndBaz", "with foo and bar", Map("foo" -> JsObject(), "bar" -> JsObject("test" -> JsString("test")), "baz" -> JsObject()))
-          param.asAdmin(Post("/v0/records", withFooAndBarAndBaz)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          val withFooAndBarAndBaz = Record("withFooAndBarAndBaz", "with foo and bar", Map("foo" -> JsObject(), "bar" -> JsObject("test" -> JsString("test")), "baz" -> JsObject()), tenantId = Some(TENANT_1))
+          param.asAdmin(Post("/v0/records", withFooAndBarAndBaz)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val withNone = Record("withNone", "with none", Map())
-          param.asAdmin(Post("/v0/records", withNone)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", withNone)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          Get("/v0/records?aspect=foo&optionalAspect=bar") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records?aspect=foo&optionalAspect=bar") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val page = responseAs[RecordsPage[Record]]
             page.records.length shouldBe 2
-            page.records.head shouldBe withFoo.copy(tenantId = Some(SINGLE_TENANT_ID))
+            page.records.head shouldBe withFoo
 
             // withFooAndBarAndBaz shouldn't include the baz aspect because it wasn't requested
-            page.records(1) shouldBe withFooAndBarAndBaz.copy(aspects = withFooAndBarAndBaz.aspects - "baz", tenantId = Some(SINGLE_TENANT_ID))
+            page.records(1) shouldBe withFooAndBarAndBaz.copy(aspects = withFooAndBarAndBaz.aspects - "baz")
+          }
+
+          Get("/v0/records?aspect=foo&optionalAspect=bar") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            val page = responseAs[RecordsPage[Record]]
+            page.records.length shouldBe 0
           }
         }
 
         it("accepts URL-encoded aspect names") { param =>
           val aspect = AspectDefinition("with space", "foo", None)
-          param.asAdmin(Post("/v0/aspects", aspect)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", aspect)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          val record = Record("whatever", "whatever", Map("with space" -> JsObject("test" -> JsString("test"))))
-          param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          val record = Record("whatever", "whatever", Map("with space" -> JsObject("test" -> JsString("test"))), tenantId = Some(TENANT_1))
+          param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          Get("/v0/records?optionalAspect=with%20space") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
-            status shouldEqual StatusCodes.OK
-            val page = responseAs[RecordsPage[Record]]
-            page.records.length shouldBe 1
-            page.records.head shouldBe record.copy(tenantId = Some(SINGLE_TENANT_ID))
-          }
-
-          Get("/v0/records?aspect=with%20space") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records?optionalAspect=with%20space") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val page = responseAs[RecordsPage[Record]]
             page.records.length shouldBe 1
-            page.records.head shouldBe record.copy(tenantId = Some(SINGLE_TENANT_ID))
+            page.records.head shouldBe record
           }
 
-          Get("/v0/records/count?aspect=with%20space") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records?aspect=with%20space") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            val page = responseAs[RecordsPage[Record]]
+            page.records.length shouldBe 1
+            page.records.head shouldBe record
+          }
+
+          Get("/v0/records/count?aspect=with%20space") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val countResponse = responseAs[CountResponse]
             countResponse.count shouldBe 1
           }
+
+          Get("/v0/records?optionalAspect=with%20space") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            val page = responseAs[RecordsPage[Record]]
+            page.records.length shouldBe 0
+          }
+
+          Get("/v0/records?aspect=with%20space") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            val page = responseAs[RecordsPage[Record]]
+            page.records.length shouldBe 0
+          }
+
+          Get("/v0/records/count?aspect=with%20space") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            val countResponse = responseAs[CountResponse]
+            countResponse.count shouldBe 0
+          }
         }
 
+        it("returns the specified aspect of the specified record of the specified tenant") { param =>
+          val aspectDefinition_1 = AspectDefinition("test", "test", None)
+          param.asAdmin(Post("/v0/aspects", aspectDefinition_1)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
+            status shouldEqual StatusCodes.OK
+          }
+          val aspectDefinition_2 = AspectDefinition("test", "test", None)
+          param.asAdmin(Post("/v0/aspects", aspectDefinition_2)) ~> addTenantIdHeader(TENANT_2) ~> param.api(Full).routes ~> check {
+            status shouldEqual StatusCodes.OK
+          }
+
+          val expected_aspect_1 = JsObject("foo_1" -> JsString("bar_1"))
+          val recordWithAspect_1 = Record("with", "with", Map("test" -> expected_aspect_1))
+          param.asAdmin(Post("/v0/records", recordWithAspect_1)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
+            status shouldEqual StatusCodes.OK
+          }
+
+          val expected_aspect_2 = JsObject("foo_2" -> JsString("bar_2"))
+          val recordWithAspect_2 = Record("with", "with", Map("test" -> expected_aspect_2))
+          param.asAdmin(Post("/v0/records", recordWithAspect_2)) ~> addTenantIdHeader(TENANT_2) ~> param.api(Full).routes ~> check {
+            status shouldEqual StatusCodes.OK
+          }
+
+          Get("/v0/records/with/aspects/test") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            val actual_aspect_1 = responseAs[JsObject]
+            actual_aspect_1 shouldBe expected_aspect_1
+          }
+
+          Get("/v0/records/with/aspects/test") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            val actual_aspect_2 = responseAs[JsObject]
+            actual_aspect_2 shouldBe expected_aspect_2
+          }
+        }
       }
 
       describe("dereference") {
         it("dereferences a single link if requested") { param =>
           val jsonSchema =
             """
-            |{
-            |    "$schema": "http://json-schema.org/hyper-schema#",
-            |    "title": "An aspect with a single link",
-            |    "type": "object",
-            |    "properties": {
-            |        "someLink": {
-            |            "title": "A link to another record.",
-            |            "type": "string",
-            |            "links": [
-            |                {
-            |                    "href": "/api/v0/registry/records/{$}",
-            |                    "rel": "item"
-            |                }
-            |            ]
-            |        }
-            |    }
-            |}
-          """.stripMargin
+              |{
+              |    "$schema": "http://json-schema.org/hyper-schema#",
+              |    "title": "An aspect with a single link",
+              |    "type": "object",
+              |    "properties": {
+              |        "someLink": {
+              |            "title": "A link to another record.",
+              |            "type": "string",
+              |            "links": [
+              |                {
+              |                    "href": "/api/v0/registry/records/{$}",
+              |                    "rel": "item"
+              |                }
+              |            ]
+              |        }
+              |    }
+              |}
+            """.stripMargin
           val aspect = AspectDefinition("withLink", "with link", Some(JsonParser(jsonSchema).asJsObject))
-          param.asAdmin(Post("/v0/aspects", aspect)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", aspect)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val source = Record("source", "source", Map("withLink" -> JsObject("someLink" -> JsString("target"))))
-          param.asAdmin(Post("/v0/records", source)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", source)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val target = Record("target", "target", Map("withLink" -> JsObject("someLink" -> JsString("source"))))
-          param.asAdmin(Post("/v0/records", target)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", target)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          Get("/v0/records/source?aspect=withLink") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
-            status shouldEqual StatusCodes.OK
-            responseAs[Record].aspects("withLink") shouldBe JsObject(
-              "someLink" -> JsString("target"))
-          }
-
-          Get("/v0/records/source?aspect=withLink&dereference=false") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records/source?aspect=withLink") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             responseAs[Record].aspects("withLink") shouldBe JsObject(
               "someLink" -> JsString("target"))
           }
 
-          Get("/v0/records/source?aspect=withLink&dereference=true") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records/source?aspect=withLink&dereference=false") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            responseAs[Record].aspects("withLink") shouldBe JsObject(
+              "someLink" -> JsString("target"))
+          }
+
+          Get("/v0/records/source?aspect=withLink&dereference=true") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             responseAs[Record].aspects("withLink") shouldBe JsObject(
               "someLink" -> JsObject(
@@ -430,66 +540,70 @@ class RecordsServiceSpec extends ApiSpec {
                   "withLink" -> JsObject(
                     "someLink" -> JsString("source")))))
           }
+
+          Get("/v0/records/source?aspect=withLink") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.NotFound
+          }
         }
 
         it("dereferences an array of links if requested") { param =>
           val jsonSchema =
             """
-            |{
-            |    "$schema": "http://json-schema.org/hyper-schema#",
-            |    "title": "An aspect with an array of links",
-            |    "type": "object",
-            |    "properties": {
-            |        "someLinks": {
-            |            "title": "Link to other records.",
-            |            "type": "array",
-            |            "items": {
-            |                "title": "A link",
-            |                "type": "string",
-            |                "links": [
-            |                    {
-            |                        "href": "/api/v0/registry/records/{$}",
-            |                        "rel": "item"
-            |                    }
-            |                ]
-            |            }
-            |        }
-            |    }
-            |}
-          """.stripMargin
+              |{
+              |    "$schema": "http://json-schema.org/hyper-schema#",
+              |    "title": "An aspect with an array of links",
+              |    "type": "object",
+              |    "properties": {
+              |        "someLinks": {
+              |            "title": "Link to other records.",
+              |            "type": "array",
+              |            "items": {
+              |                "title": "A link",
+              |                "type": "string",
+              |                "links": [
+              |                    {
+              |                        "href": "/api/v0/registry/records/{$}",
+              |                        "rel": "item"
+              |                    }
+              |                ]
+              |            }
+              |        }
+              |    }
+              |}
+            """.stripMargin
           val aspect = AspectDefinition("withLinks", "with links", Some(JsonParser(jsonSchema).asJsObject))
-          param.asAdmin(Post("/v0/aspects", aspect)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", aspect)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val source = Record("source", "source", Map("withLinks" -> JsObject("someLinks" -> JsArray(JsString("target"), JsString("anotherTarget")))))
-          param.asAdmin(Post("/v0/records", source)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", source)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val target = Record("target", "target", Map("withLinks" -> JsObject("someLinks" -> JsArray(JsString("source")))))
-          param.asAdmin(Post("/v0/records", target)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", target)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val anotherTarget = Record("anotherTarget", "anotherTarget", Map("withLinks" -> JsObject("someLinks" -> JsArray(JsString("source")))))
-          param.asAdmin(Post("/v0/records", anotherTarget)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", anotherTarget)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          Get("/v0/records/source?aspect=withLinks") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
-            status shouldEqual StatusCodes.OK
-            responseAs[Record].aspects("withLinks") shouldBe JsObject(
-              "someLinks" -> JsArray(JsString("target"), JsString("anotherTarget")))
-          }
-
-          Get("/v0/records/source?aspect=withLinks&dereference=false") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records/source?aspect=withLinks") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             responseAs[Record].aspects("withLinks") shouldBe JsObject(
               "someLinks" -> JsArray(JsString("target"), JsString("anotherTarget")))
           }
 
-          Get("/v0/records/source?aspect=withLinks&dereference=true") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records/source?aspect=withLinks&dereference=false") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            responseAs[Record].aspects("withLinks") shouldBe JsObject(
+              "someLinks" -> JsArray(JsString("target"), JsString("anotherTarget")))
+          }
+
+          Get("/v0/records/source?aspect=withLinks&dereference=true") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             responseAs[Record].aspects("withLinks") shouldBe JsObject(
               "someLinks" -> JsArray(
@@ -511,39 +625,39 @@ class RecordsServiceSpec extends ApiSpec {
         it("should not excludes linking aspects when there are no links and dereference=true") { param =>
           val jsonSchema =
             """
-            |{
-            |    "$schema": "http://json-schema.org/hyper-schema#",
-            |    "title": "An aspect with an array of links",
-            |    "type": "object",
-            |    "properties": {
-            |        "someLinks": {
-            |            "title": "Link to other records.",
-            |            "type": "array",
-            |            "items": {
-            |                "title": "A link",
-            |                "type": "string",
-            |                "links": [
-            |                    {
-            |                        "href": "/api/v0/registry/records/{$}",
-            |                        "rel": "item"
-            |                    }
-            |                ]
-            |            }
-            |        }
-            |    }
-            |}
-          """.stripMargin
+              |{
+              |    "$schema": "http://json-schema.org/hyper-schema#",
+              |    "title": "An aspect with an array of links",
+              |    "type": "object",
+              |    "properties": {
+              |        "someLinks": {
+              |            "title": "Link to other records.",
+              |            "type": "array",
+              |            "items": {
+              |                "title": "A link",
+              |                "type": "string",
+              |                "links": [
+              |                    {
+              |                        "href": "/api/v0/registry/records/{$}",
+              |                        "rel": "item"
+              |                    }
+              |                ]
+              |            }
+              |        }
+              |    }
+              |}
+            """.stripMargin
           val aspect = AspectDefinition("withLinks", "with links", Some(JsonParser(jsonSchema).asJsObject))
-          param.asAdmin(Post("/v0/aspects", aspect)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", aspect)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val source = Record("source", "source", Map("withLinks" -> JsObject("someLinks" -> JsArray())))
-          param.asAdmin(Post("/v0/records", source)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", source)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          Get("/v0/records/source?aspect=withLinks&dereference=true") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records/source?aspect=withLinks&dereference=true") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             responseAs[Record].aspects("withLinks") shouldBe JsObject(
               "someLinks" -> JsArray())
@@ -554,34 +668,34 @@ class RecordsServiceSpec extends ApiSpec {
       describe("querying by aspect value") {
         it("works for shallow paths") { param =>
           val aspect = AspectDefinition("exampleAspect", "exampleAspect", None)
-          param.asAdmin(Post("/v0/aspects", aspect)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", aspect)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          val recordWithValue1 = Record("withValue1", "withValue1", Map("exampleAspect" -> JsObject("value" -> JsString("correct"))))
-          param.asAdmin(Post("/v0/records", recordWithValue1)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          val recordWithValue1 = Record("withValue1", "withValue1", Map("exampleAspect" -> JsObject("value" -> JsString("correct"))), tenantId = Some(TENANT_1))
+          param.asAdmin(Post("/v0/records", recordWithValue1)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          val recordWithValue2 = Record("withValue2", "withValue2", Map("exampleAspect" -> JsObject("value" -> JsString("correct"))))
-          param.asAdmin(Post("/v0/records", recordWithValue2)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          val recordWithValue2 = Record("withValue2", "withValue2", Map("exampleAspect" -> JsObject("value" -> JsString("correct"))), tenantId = Some(TENANT_1))
+          param.asAdmin(Post("/v0/records", recordWithValue2)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val recordWithoutValue = Record("withoutValue", "withoutValue", Map("exampleAspect" -> JsObject("value" -> JsString("incorrect"))))
-          param.asAdmin(Post("/v0/records", recordWithoutValue)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", recordWithoutValue)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          Get("/v0/records?aspectQuery=exampleAspect.value:correct&aspect=exampleAspect") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records?aspectQuery=exampleAspect.value:correct&aspect=exampleAspect") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val page = responseAs[RecordsPage[Record]]
             page.records.length shouldBe 2
-            page.records.head shouldBe recordWithValue1.copy(tenantId = Some(SINGLE_TENANT_ID))
-            page.records(1) shouldBe recordWithValue2.copy(tenantId = Some(SINGLE_TENANT_ID))
+            page.records.head shouldBe recordWithValue1
+            page.records(1) shouldBe recordWithValue2
           }
 
-          Get("/v0/records/count?aspectQuery=exampleAspect.value:correct&aspect=exampleAspect") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records/count?aspectQuery=exampleAspect.value:correct&aspect=exampleAspect") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val countResponse = responseAs[CountResponse]
             countResponse.count shouldBe 2
@@ -590,26 +704,26 @@ class RecordsServiceSpec extends ApiSpec {
 
         it("works without specifying the aspect in aspects or optionalAspects") { param =>
           val aspect = AspectDefinition("exampleAspect", "exampleAspect", None)
-          param.asAdmin(Post("/v0/aspects", aspect)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", aspect)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val recordWithValue1 = Record("withValue1", "withValue1", Map("exampleAspect" -> JsObject("value" -> JsString("correct"))))
-          param.asAdmin(Post("/v0/records", recordWithValue1)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", recordWithValue1)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val recordWithValue2 = Record("withValue2", "withValue2", Map("exampleAspect" -> JsObject("value" -> JsString("correct"))))
-          param.asAdmin(Post("/v0/records", recordWithValue2)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", recordWithValue2)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val recordWithoutValue = Record("withoutValue", "withoutValue", Map("exampleAspect" -> JsObject("value" -> JsString("incorrect"))))
-          param.asAdmin(Post("/v0/records", recordWithoutValue)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", recordWithoutValue)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          Get("/v0/records?aspectQuery=exampleAspect.value:correct") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records?aspectQuery=exampleAspect.value:correct") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val page = responseAs[RecordsPage[Record]]
             page.records.length shouldBe 2
@@ -617,7 +731,7 @@ class RecordsServiceSpec extends ApiSpec {
             page.records(1).id shouldBe "withValue2"
           }
 
-          Get("/v0/records/count?aspectQuery=exampleAspect.value:correct") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records/count?aspectQuery=exampleAspect.value:correct") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val countResponse = responseAs[CountResponse]
             countResponse.count shouldBe 2
@@ -626,34 +740,34 @@ class RecordsServiceSpec extends ApiSpec {
 
         it("works for deep paths") { param =>
           val aspect = AspectDefinition("exampleAspect", "exampleAspect", None)
-          param.asAdmin(Post("/v0/aspects", aspect)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", aspect)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          val recordWithValue1 = Record("withValue1", "withValue1", Map("exampleAspect" -> JsObject("object" -> JsObject("value" -> JsString("correct")))))
-          param.asAdmin(Post("/v0/records", recordWithValue1)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          val recordWithValue1 = Record("withValue1", "withValue1", Map("exampleAspect" -> JsObject("object" -> JsObject("value" -> JsString("correct")))), tenantId = Some(TENANT_1))
+          param.asAdmin(Post("/v0/records", recordWithValue1)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          val recordWithValue2 = Record("withValue2", "withValue2", Map("exampleAspect" -> JsObject("object" -> JsObject("value" -> JsString("correct")))))
-          param.asAdmin(Post("/v0/records", recordWithValue2)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          val recordWithValue2 = Record("withValue2", "withValue2", Map("exampleAspect" -> JsObject("object" -> JsObject("value" -> JsString("correct")))), tenantId = Some(TENANT_1))
+          param.asAdmin(Post("/v0/records", recordWithValue2)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val recordWithoutValue = Record("withoutValue", "withoutValue", Map("exampleAspect" -> JsObject("object" -> JsObject("value" -> JsString("incorrect")))))
-          param.asAdmin(Post("/v0/records", recordWithoutValue)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", recordWithoutValue)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          Get("/v0/records?aspectQuery=exampleAspect.object.value:correct&aspect=exampleAspect") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records?aspectQuery=exampleAspect.object.value:correct&aspect=exampleAspect") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val page = responseAs[RecordsPage[Record]]
             page.records.length shouldBe 2
-            page.records.head shouldBe recordWithValue1.copy(tenantId = Some(SINGLE_TENANT_ID))
-            page.records(1) shouldBe recordWithValue2.copy(tenantId = Some(SINGLE_TENANT_ID))
+            page.records.head shouldBe recordWithValue1
+            page.records(1) shouldBe recordWithValue2
           }
 
-          Get("/v0/records/count?aspectQuery=exampleAspect.object.value:correct&aspect=exampleAspect") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records/count?aspectQuery=exampleAspect.object.value:correct&aspect=exampleAspect") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val countResponse = responseAs[CountResponse]
             countResponse.count shouldBe 2
@@ -662,34 +776,34 @@ class RecordsServiceSpec extends ApiSpec {
 
         it("works as AND when multiple queries specified") { param =>
           val aspect = AspectDefinition("exampleAspect", "exampleAspect", None)
-          param.asAdmin(Post("/v0/aspects", aspect)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", aspect)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          val recordWithValue1 = Record("withValue1", "withValue1", Map("exampleAspect" -> JsObject("value" -> JsString("correct"), "otherValue" -> JsString("alsoCorrect"))))
-          param.asAdmin(Post("/v0/records", recordWithValue1)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          val recordWithValue1 = Record("withValue1", "withValue1", Map("exampleAspect" -> JsObject("value" -> JsString("correct"), "otherValue" -> JsString("alsoCorrect"))), tenantId = Some(TENANT_1))
+          param.asAdmin(Post("/v0/records", recordWithValue1)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          val recordWithValue2 = Record("withValue2", "withValue2", Map("exampleAspect" -> JsObject("value" -> JsString("correct"), "otherValue" -> JsString("alsoCorrect"))))
-          param.asAdmin(Post("/v0/records", recordWithValue2)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          val recordWithValue2 = Record("withValue2", "withValue2", Map("exampleAspect" -> JsObject("value" -> JsString("correct"), "otherValue" -> JsString("alsoCorrect"))), tenantId = Some(TENANT_1))
+          param.asAdmin(Post("/v0/records", recordWithValue2)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val recordWithoutValue = Record("withoutValue", "withoutValue", Map("exampleAspect" -> JsObject("value" -> JsString("correct"))))
-          param.asAdmin(Post("/v0/records", recordWithoutValue)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", recordWithoutValue)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          Get("/v0/records?aspectQuery=exampleAspect.value:correct&aspectQuery=exampleAspect.otherValue:alsoCorrect&aspect=exampleAspect") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records?aspectQuery=exampleAspect.value:correct&aspectQuery=exampleAspect.otherValue:alsoCorrect&aspect=exampleAspect") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val page = responseAs[RecordsPage[Record]]
             page.records.length shouldBe 2
-            page.records.head shouldBe recordWithValue1.copy(tenantId = Some(SINGLE_TENANT_ID))
-            page.records(1) shouldBe recordWithValue2.copy(tenantId = Some(SINGLE_TENANT_ID))
+            page.records.head shouldBe recordWithValue1
+            page.records(1) shouldBe recordWithValue2
           }
 
-          Get("/v0/records/count?aspectQuery=exampleAspect.value:correct&aspectQuery=exampleAspect.otherValue:alsoCorrect&aspect=exampleAspect") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records/count?aspectQuery=exampleAspect.value:correct&aspectQuery=exampleAspect.otherValue:alsoCorrect&aspect=exampleAspect") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val countResponse = responseAs[CountResponse]
             countResponse.count shouldBe 2
@@ -698,28 +812,28 @@ class RecordsServiceSpec extends ApiSpec {
 
         it("allows url encoded paths and values") { param =>
           val aspect = AspectDefinition("example Aspect", "example Aspect", None)
-          param.asAdmin(Post("/v0/aspects", aspect)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", aspect)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          val recordWithValue = Record("withValue", "withValue", Map("example Aspect" -> JsObject("&value" -> JsString("/correct"))))
-          param.asAdmin(Post("/v0/records", recordWithValue)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          val recordWithValue = Record("withValue", "withValue", Map("example Aspect" -> JsObject("&value" -> JsString("/correct"))), tenantId = Some(TENANT_1))
+          param.asAdmin(Post("/v0/records", recordWithValue)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val recordWithoutValue = Record("withoutValue", "withoutValue", Map("example Aspect" -> JsObject("value" -> JsString("incorrect"))))
-          param.asAdmin(Post("/v0/records", recordWithoutValue)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+          param.asAdmin(Post("/v0/records", recordWithoutValue)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          Get("/v0/records?aspectQuery=example%20Aspect.%26value:%2Fcorrect&aspect=example%20Aspect") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records?aspectQuery=example%20Aspect.%26value:%2Fcorrect&aspect=example%20Aspect") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val page = responseAs[RecordsPage[Record]]
             page.records.length shouldBe 1
-            page.records.head shouldBe recordWithValue.copy(tenantId = Some(SINGLE_TENANT_ID))
+            page.records.head shouldBe recordWithValue
           }
 
-          Get("/v0/records/count?aspectQuery=example%20Aspect.%26value:%2Fcorrect&aspect=example%20Aspect") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records/count?aspectQuery=example%20Aspect.%26value:%2Fcorrect&aspect=example%20Aspect") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val countResponse = responseAs[CountResponse]
             countResponse.count shouldBe 1
@@ -739,18 +853,18 @@ class RecordsServiceSpec extends ApiSpec {
         def pagingTests(path: String) {
           it("honors the limit parameter") { param =>
             val aspectDefinition = AspectDefinition("test", "test", None)
-            param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+            param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
               status shouldEqual StatusCodes.OK
             }
 
             for (i <- 1 to 5) {
               val record = Record(i.toString, i.toString, Map("test" -> JsObject("value" -> JsNumber(i))))
-              param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+              param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
                 status shouldEqual StatusCodes.OK
               }
             }
 
-            Get(s"/v0/records${path}limit=2") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+            Get(s"/v0/records${path}limit=2") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
               status shouldEqual StatusCodes.OK
               val page = responseAs[RecordsPage[RecordType]]
               page.records.length shouldBe 2
@@ -761,18 +875,18 @@ class RecordsServiceSpec extends ApiSpec {
 
           it("honors the start parameter") { param =>
             val aspectDefinition = AspectDefinition("test", "test", None)
-            param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+            param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
               status shouldEqual StatusCodes.OK
             }
 
             for (i <- 1 to 5) {
               val record = Record(i.toString, i.toString, Map("test" -> JsObject("value" -> JsNumber(i))))
-              param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+              param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
                 status shouldEqual StatusCodes.OK
               }
             }
 
-            Get(s"/v0/records${path}start=3&limit=2") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+            Get(s"/v0/records${path}start=3&limit=2") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
               status shouldEqual StatusCodes.OK
               val page = responseAs[RecordsPage[RecordType]]
               page.records.length shouldBe 2
@@ -783,18 +897,18 @@ class RecordsServiceSpec extends ApiSpec {
 
           it("pageTokens can be used to page through results") { param =>
             val aspectDefinition = AspectDefinition("test", "test", None)
-            param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+            param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
               status shouldEqual StatusCodes.OK
             }
 
             for (i <- 1 to 5) {
               val record = Record(i.toString, i.toString, Map("test" -> JsObject("value" -> JsNumber(i))))
-              param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+              param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
                 status shouldEqual StatusCodes.OK
               }
             }
 
-            var currentPage = Get(s"/v0/records${path}limit=2") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+            var currentPage = Get(s"/v0/records${path}limit=2") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
               status shouldEqual StatusCodes.OK
               val page = responseAs[RecordsPage[RecordType]]
               page.records.length shouldBe 2
@@ -804,7 +918,7 @@ class RecordsServiceSpec extends ApiSpec {
             }
 
             currentPage =
-              Get(s"/v0/records${path}pageToken=${currentPage.nextPageToken.get}&limit=2") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+              Get(s"/v0/records${path}pageToken=${currentPage.nextPageToken.get}&limit=2") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
                 status shouldEqual StatusCodes.OK
                 val page = responseAs[RecordsPage[RecordType]]
                 page.records.length shouldBe 2
@@ -813,7 +927,7 @@ class RecordsServiceSpec extends ApiSpec {
                 page
               }
 
-            currentPage = Get(s"/v0/records${path}pageToken=${currentPage.nextPageToken.get}&limit=2") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+            currentPage = Get(s"/v0/records${path}pageToken=${currentPage.nextPageToken.get}&limit=2") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
               status shouldEqual StatusCodes.OK
               val page = responseAs[RecordsPage[RecordType]]
               page.records.length shouldBe 1
@@ -824,53 +938,53 @@ class RecordsServiceSpec extends ApiSpec {
 
           it("provides hasMore correctly") { param =>
             val aspectDefinition = AspectDefinition("test", "test", None)
-            param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+            param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
               status shouldEqual StatusCodes.OK
             }
 
             for (i <- 1 to 5) {
               val record = Record(i.toString, i.toString, Map("test" -> JsObject("value" -> JsNumber(i))))
-              param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+              param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
                 status shouldEqual StatusCodes.OK
               }
             }
 
-            Get(s"/v0/records${path}start=0&limit=4") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+            Get(s"/v0/records${path}start=0&limit=4") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
               status shouldEqual StatusCodes.OK
               val page = responseAs[RecordsPage[RecordType]]
               page.hasMore shouldBe true
               page.nextPageToken.isDefined shouldBe true
             }
 
-            Get(s"/v0/records${path}start=0&limit=5") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+            Get(s"/v0/records${path}start=0&limit=5") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
               status shouldEqual StatusCodes.OK
               val page = responseAs[RecordsPage[RecordType]]
               page.hasMore shouldBe false
               page.nextPageToken.isDefined shouldBe false
             }
 
-            Get(s"/v0/records${path}start=0&limit=6") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+            Get(s"/v0/records${path}start=0&limit=6") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
               status shouldEqual StatusCodes.OK
               val page = responseAs[RecordsPage[RecordType]]
               page.hasMore shouldBe false
               page.nextPageToken.isDefined shouldBe false
             }
 
-            Get(s"/v0/records${path}start=3&limit=1") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+            Get(s"/v0/records${path}start=3&limit=1") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
               status shouldEqual StatusCodes.OK
               val page = responseAs[RecordsPage[RecordType]]
               page.hasMore shouldBe true
               page.nextPageToken.isDefined shouldBe true
             }
 
-            Get(s"/v0/records${path}start=4&limit=1") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+            Get(s"/v0/records${path}start=4&limit=1") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
               status shouldEqual StatusCodes.OK
               val page = responseAs[RecordsPage[RecordType]]
               page.hasMore shouldBe false
               page.nextPageToken.isDefined shouldBe false
             }
 
-            Get(s"/v0/records${path}start=5&limit=1") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+            Get(s"/v0/records${path}start=5&limit=1") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
               status shouldEqual StatusCodes.OK
               val page = responseAs[RecordsPage[RecordType]]
               page.hasMore shouldBe false
@@ -892,7 +1006,7 @@ class RecordsServiceSpec extends ApiSpec {
                 it(s"for pageSize $pageSize and recordCount $recordCount") { param =>
                   // Add an aspect for our records
                   val aspectDefinition = AspectDefinition("test", "test", None)
-                  param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+                  param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
                     status shouldEqual StatusCodes.OK
                   }
 
@@ -900,14 +1014,14 @@ class RecordsServiceSpec extends ApiSpec {
                   if (recordCount > 0) {
                     for (i <- 1 to recordCount) {
                       val record = Record(i.toString, i.toString, Map("test" -> JsObject("value" -> JsNumber(i))))
-                      param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+                      param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
                         status shouldEqual StatusCodes.OK
                       }
                     }
                   }
 
                   // Get the page tokens
-                  Get(s"/v0/records/pagetokens?limit=$pageSize") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+                  Get(s"/v0/records/pagetokens?limit=$pageSize") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
                     status shouldEqual StatusCodes.OK
                     val pageTokens = responseAs[List[String]]
                     pageTokens.length shouldBe recordCount / Math.max(1, pageSize) + 1
@@ -916,7 +1030,7 @@ class RecordsServiceSpec extends ApiSpec {
                     for (pageIndex <- pageTokens.indices) {
                       val token = pageTokens(pageIndex)
 
-                      Get(s"/v0/records?pageToken=$token&limit=$pageSize") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+                      Get(s"/v0/records?pageToken=$token&limit=$pageSize") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
                         status shouldEqual StatusCodes.OK
                         val page = responseAs[RecordsPage[Record]]
 
@@ -937,14 +1051,14 @@ class RecordsServiceSpec extends ApiSpec {
                   // Insert two aspects
                   for (aspectNumber <- 1 to 2) {
                     val aspectDefinition = AspectDefinition(aspectNumber.toString, aspectNumber.toString, None)
-                    param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+                    param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
                       status shouldEqual StatusCodes.OK
 
                       // Insert $recordCount records for each aspect
                       if (recordCount > 0) {
                         for (i <- 1 to recordCount) {
                           val record = Record(aspectNumber + i.toString, i.toString, Map(aspectNumber.toString -> JsObject("value" -> JsNumber(i))))
-                          param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+                          param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
                             status shouldEqual StatusCodes.OK
                           }
                         }
@@ -954,7 +1068,7 @@ class RecordsServiceSpec extends ApiSpec {
 
                   // Filter by each aspect and make sure that the tokens returned match up with their respective GET /records results.
                   for (aspectNumber <- 1 to 2) {
-                    Get(s"/v0/records/pagetokens?limit=$pageSize&aspect=$aspectNumber") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+                    Get(s"/v0/records/pagetokens?limit=$pageSize&aspect=$aspectNumber") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
                       status shouldEqual StatusCodes.OK
                       val pageTokens = responseAs[List[String]]
                       pageTokens.length shouldEqual recordCount / Math.max(1, pageSize) + 1
@@ -962,7 +1076,7 @@ class RecordsServiceSpec extends ApiSpec {
                       for (pageIndex <- pageTokens.indices) {
                         val token = pageTokens(pageIndex)
 
-                        Get(s"/v0/records?pageToken=$token&limit=$pageSize&aspect=$aspectNumber") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+                        Get(s"/v0/records?pageToken=$token&limit=$pageSize&aspect=$aspectNumber") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
                           status shouldEqual StatusCodes.OK
                           val page = responseAs[RecordsPage[Record]]
 
@@ -985,7 +1099,7 @@ class RecordsServiceSpec extends ApiSpec {
                   val aspectIds = for (aspectNumber <- 1 to 2) yield aspectNumber.toString
                   aspectIds.foreach { aspectNumber =>
                     val aspectDefinition = AspectDefinition(aspectNumber, aspectNumber, None)
-                    param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+                    param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
                       status shouldEqual StatusCodes.OK
                     }
                   }
@@ -995,7 +1109,7 @@ class RecordsServiceSpec extends ApiSpec {
                     for (i <- 1 to recordCount) {
                       val aspectValues = aspectIds.map(id => id -> JsObject("value" -> JsNumber(i)))
                       val record = Record(i.toString, i.toString, aspectValues.toMap)
-                      param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+                      param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
                         status shouldEqual StatusCodes.OK
                       }
                     }
@@ -1005,7 +1119,7 @@ class RecordsServiceSpec extends ApiSpec {
                   aspectIds.foreach { aspectId =>
                     for (i <- 1 to recordCount) {
                       val record = Record(aspectId + i.toString, i.toString, Map(aspectId -> JsObject("value" -> JsNumber(i))))
-                      param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(Full).routes ~> check {
+                      param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(Full).routes ~> check {
                         status shouldEqual StatusCodes.OK
                       }
                     }
@@ -1013,7 +1127,7 @@ class RecordsServiceSpec extends ApiSpec {
 
                   // Check that pagetokens while filtering for both aspects only returns the records with both, not the
                   // records with one or the other.
-                  Get(s"/v0/records/pagetokens?limit=$pageSize&${aspectIds.map(id => "aspect=" + id).mkString("&")}") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+                  Get(s"/v0/records/pagetokens?limit=$pageSize&${aspectIds.map(id => "aspect=" + id).mkString("&")}") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
                     status shouldEqual StatusCodes.OK
                     val pageTokens = responseAs[List[String]]
                     pageTokens.length shouldEqual recordCount / Math.max(1, pageSize) + 1
@@ -1021,7 +1135,7 @@ class RecordsServiceSpec extends ApiSpec {
                     for (pageIndex <- pageTokens.indices) {
                       val token = pageTokens(pageIndex)
 
-                      Get(s"/v0/records?pageToken=$token&limit=$pageSize&${aspectIds.map(id => "aspect=" + id).mkString("&")}") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+                      Get(s"/v0/records?pageToken=$token&limit=$pageSize&${aspectIds.map(id => "aspect=" + id).mkString("&")}") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
                         status shouldEqual StatusCodes.OK
                         val page = responseAs[RecordsPage[Record]]
 
@@ -1043,58 +1157,103 @@ class RecordsServiceSpec extends ApiSpec {
     describe("POST") {
       it("can add a new record") { param =>
         val record = Record("testId", "testName", Map(), Some("tag"))
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual record.copy(tenantId = Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual record.copy(tenantId = Some(TENANT_1))
         }
 
-        Get("/v0/records/testId") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        Get("/v0/records/testId") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
 
           val recordRes = responseAs[Record]
-          recordRes shouldEqual record.copy(tenantId = Some(SINGLE_TENANT_ID))
+          recordRes shouldEqual record.copy(tenantId = Some(TENANT_1))
+        }
+
+        Get("/v0/records/testId") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.NotFound
+        }
+      }
+
+      it("can add two new records with the same record IDs by different tenants") { param =>
+        val record_1 = Record("aRecordId", "a default tenant", Map(), Some("tag"))
+        param.asAdmin(Post("/v0/records", record_1)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.OK
+          responseAs[Record] shouldEqual record_1.copy(tenantId = Some(TENANT_1))
+        }
+
+
+        val record_2 = Record("aRecordId", "a new tenant", Map(), Some("tag"))
+        param.asAdmin(Post("/v0/records", record_2)) ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.OK
+          responseAs[Record] shouldEqual record_2.copy(tenantId = Some(TENANT_2))
+        }
+
+        Get("/v0/records/aRecordId") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.OK
+
+          val recordRes = responseAs[Record]
+          recordRes shouldEqual record_1.copy(tenantId = Some(TENANT_1))
+        }
+
+        Get("/v0/records/aRecordId") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.OK
+
+          val recordRes = responseAs[Record]
+          recordRes shouldEqual record_2.copy(tenantId = Some(TENANT_2))
         }
       }
 
       it("sets sourcetag to NULL by default") { param =>
         val record = Record("testId", "testName", Map(), None)
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual record.copy(tenantId = Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual record.copy(tenantId = Some(TENANT_1))
         }
 
-        Get("/v0/records/testId") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        Get("/v0/records/testId") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
 
           val recordRes = responseAs[Record]
-          recordRes shouldEqual record.copy(tenantId = Some(SINGLE_TENANT_ID))
+          recordRes shouldEqual record.copy(tenantId = Some(TENANT_1))
+        }
+
+        Get("/v0/records/testId") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.NotFound
         }
       }
 
       it("supports invalid URL characters in ID") { param =>
         val record = Record("in valid", "testName", Map())
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual record.copy(tenantId = Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual record.copy(tenantId = Some(TENANT_1))
         }
 
-        Get("/v0/records/in%20valid") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        Get("/v0/records/in%20valid") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual Record("in valid", "testName", Map(), None, Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual Record("in valid", "testName", Map(), tenantId = Some(TENANT_1))
+        }
+
+        Get("/v0/records/in%20valid") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.NotFound
         }
       }
 
       it("returns 400 if a record with the given ID already exists") { param =>
         val record = Record("testId", "testName", Map())
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual record.copy(tenantId = Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual record.copy(tenantId = Some(TENANT_1))
         }
 
         val updated = record.copy(name = "foo")
-        param.asAdmin(Post("/v0/records", updated)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", updated)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.BadRequest
           responseAs[BadRequest].message should include("already exists")
+        }
+
+        param.asAdmin(Post("/v0/records", updated)) ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.OK
         }
       }
 
@@ -1107,61 +1266,97 @@ class RecordsServiceSpec extends ApiSpec {
     describe("PUT") {
       it("can add a new record") { param =>
         val record = Record("testId", "testName", Map())
-        param.asAdmin(Put("/v0/records/testId", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Put("/v0/records/testId", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual record.copy(tenantId = Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual record.copy( tenantId = Some(TENANT_1))
         }
 
-        Get("/v0/records") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        Get("/v0/records") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
 
           val recordsPage = responseAs[RecordsPage[Record]]
           recordsPage.records.length shouldEqual 1
-          recordsPage.records.head shouldEqual Record("testId", "testName", Map(), None, Some(SINGLE_TENANT_ID))
+          recordsPage.records.head shouldEqual record.copy( tenantId = Some(TENANT_1))
+        }
+
+        Get("/v0/records") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.OK
+          val recordsPage = responseAs[RecordsPage[Record]]
+          recordsPage.records.length shouldEqual 0
         }
       }
 
       it("can update an existing record") { param =>
-        val record = Record("testId", "testName", Map())
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        val record = Record("testId", "testName", Map(), tenantId = Some(TENANT_1))
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val newRecord = record.copy(name = "newName")
-        param.asAdmin(Put("/v0/records/testId", newRecord)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Put("/v0/records/testId", newRecord)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual newRecord.copy(tenantId = Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual newRecord.copy(tenantId = Some(TENANT_1))
         }
 
-        Get("/v0/records/testId") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Put("/v0/records/testId", newRecord)) ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual Record("testId", "newName", Map(), None, Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual newRecord.copy(tenantId = Some(TENANT_2))
+        }
+
+        Get("/v0/records/testId") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.OK
+          responseAs[Record] shouldEqual Record("testId", "newName", Map(), tenantId = Some(TENANT_1))
+        }
+
+        Get("/v0/records/testId") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.OK
+          responseAs[Record] shouldEqual Record("testId", "newName", Map(), tenantId = Some(TENANT_2))
         }
       }
 
       it("updates the sourcetag of an otherwise identical record without generating events") { param =>
         val record = Record("testId", "testName", Map(), Some("tag1"))
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
-        Get(s"/v0/records/${record.id}/history") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        Get(s"/v0/records/${record.id}/history") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
           responseAs[EventsPage].events.length shouldEqual 1
         }
 
-        val newRecord = record.copy(sourceTag = Some("tag2"))
-        param.asAdmin(Put("/v0/records/testId", newRecord)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        Get(s"/v0/records/${record.id}/history") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual newRecord.copy(tenantId = Some(SINGLE_TENANT_ID))
+          responseAs[EventsPage].events.length shouldEqual 0
         }
 
-        Get("/v0/records/testId") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        val newRecord = record.copy(sourceTag = Some("tag2"))
+        param.asAdmin(Put("/v0/records/testId", newRecord)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.OK
+          responseAs[Record] shouldEqual newRecord.copy(tenantId = Some(TENANT_1))
+        }
+
+        param.asAdmin(Put("/v0/records/testId", newRecord)) ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.OK
+          responseAs[Record] shouldEqual newRecord.copy(tenantId = Some(TENANT_2))
+        }
+
+        Get("/v0/records/testId") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
           responseAs[Record].sourceTag shouldEqual Some("tag2")
         }
 
-        Get(s"/v0/records/${record.id}/history") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        Get(s"/v0/records/${record.id}/history") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.OK
+          responseAs[EventsPage].events.length shouldEqual 1
+        }
+
+        Get("/v0/records/testId") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.OK
+          responseAs[Record] shouldEqual newRecord.copy(tenantId = Some(TENANT_2))
+        }
+
+        Get(s"/v0/records/${record.id}/history") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
           responseAs[EventsPage].events.length shouldEqual 1
         }
@@ -1169,99 +1364,103 @@ class RecordsServiceSpec extends ApiSpec {
 
       it("cannot change the ID of an existing record") { param =>
         val record = Record("testId", "testName", Map())
-        param.asAdmin(Put("/v0/records/testId", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Put("/v0/records/testId", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual record.copy(tenantId = Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual record.copy(tenantId = Some(TENANT_1))
         }
 
         val updated = record.copy(id = "foo")
-        param.asAdmin(Put("/v0/records/testId", updated)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Put("/v0/records/testId", updated)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.BadRequest
           responseAs[BadRequest].message should include("does not match the record")
+        }
+
+        param.asAdmin(Put("/v0/records/testId", updated)) ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.BadRequest
         }
       }
 
       it("supports invalid URL characters in ID") { param =>
         val record = Record("in valid", "testName", Map())
-        param.asAdmin(Put("/v0/records/in%20valid", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Put("/v0/records/in%20valid", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual record.copy(tenantId = Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual record.copy(tenantId = Some(TENANT_1))
         }
 
-        Get("/v0/records/in%20valid") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        Get("/v0/records/in%20valid") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual Record("in valid", "testName", Map(), None, Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual record.copy(tenantId = Some(TENANT_1))
         }
       }
 
       it("can add an aspect") { param =>
         val aspectDefinition = AspectDefinition("test", "test", None)
-        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val record = Record("testId", "testName", Map())
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val updated = record.copy(aspects = Map("test" -> JsObject()))
-        param.asAdmin(Put("/v0/records/testId", updated)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Put("/v0/records/testId", updated)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual updated.copy(tenantId = Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual updated.copy(tenantId = Some(TENANT_1))
         }
 
-        Get("/v0/records/testId?aspect=test") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        Get("/v0/records/testId?aspect=test") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual updated.copy(tenantId = Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual updated.copy(tenantId = Some(TENANT_1))
         }
       }
 
       it("can modify an aspect") { param =>
         val aspectDefinition = AspectDefinition("test", "test", None)
-        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val record = Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("bar"))))
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val updated = record.copy(aspects = Map("test" -> JsObject("foo" -> JsString("baz"))))
-        param.asAdmin(Put("/v0/records/testId", updated)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Put("/v0/records/testId", updated)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual updated.copy(tenantId = Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual updated.copy(tenantId = Some(TENANT_1))
         }
 
-        Get("/v0/records/testId?aspect=test") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        Get("/v0/records/testId?aspect=test") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual updated.copy(tenantId = Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual updated.copy(tenantId = Some(TENANT_1))
         }
       }
 
       it("does not remove aspects simply because they're missing from the PUT payload") { param =>
         val aspectDefinition = AspectDefinition("test", "test", None)
-        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val record = Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("bar"))))
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         // TODO: the PUT should return the real record, not just echo back what the user provided.
         //       i.e. the aspects should be included.  I think.
         val updated = record.copy(aspects = Map())
-        param.asAdmin(Put("/v0/records/testId", updated)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Put("/v0/records/testId", updated)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual updated.copy(tenantId = Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual updated.copy(tenantId = Some(TENANT_1))
         }
 
-        Get("/v0/records/testId?aspect=test") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        Get("/v0/records/testId?aspect=test") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual record.copy(tenantId = Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual record.copy(tenantId = Some(TENANT_1))
         }
       }
 
@@ -1274,7 +1473,7 @@ class RecordsServiceSpec extends ApiSpec {
     describe("PATCH") {
       it("returns an error when the record does not exist") { param =>
         val patch = JsonPatch()
-        param.asAdmin(Patch("/v0/records/doesnotexist", patch)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/doesnotexist", patch)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.BadRequest
           responseAs[BadRequest].message should include("exists")
           responseAs[BadRequest].message should include("ID")
@@ -1283,30 +1482,38 @@ class RecordsServiceSpec extends ApiSpec {
 
       it("can modify a record's name") { param =>
         val record = Record("testId", "testName", Map())
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val patch = JsonPatch(Replace(Pointer.root / "name", JsString("foo")))
-        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual Record("testId", "foo", Map(), None, Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual Record("testId", "foo", Map(), tenantId = Some(TENANT_1))
         }
 
-        Get("/v0/records/testId") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.BadRequest
+        }
+
+        Get("/v0/records/testId") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual Record("testId", "foo", Map(), None, Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual Record("testId", "foo", Map(), tenantId = Some(TENANT_1))
+        }
+
+        Get("/v0/records/testId") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.NotFound
         }
       }
 
       it("cannot modify a record's ID") { param =>
         val record = Record("testId", "testName", Map())
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val patch = JsonPatch(Replace(Pointer.root / "id", JsString("foo")))
-        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.BadRequest
           responseAs[BadRequest].message should include("ID")
         }
@@ -1314,184 +1521,244 @@ class RecordsServiceSpec extends ApiSpec {
 
       it("can add an aspect") { param =>
         val aspectDefinition = AspectDefinition("test", "test", None)
-        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val record = Record("testId", "testName", Map())
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val patch = JsonPatch(Add(Pointer.root / "aspects" / "test", JsObject()))
-        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject()), None, Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject()), tenantId = Some(TENANT_1))
         }
 
-        Get("/v0/records/testId?aspect=test") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.BadRequest
+        }
+
+        Get("/v0/records/testId?aspect=test") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject()), None, Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject()), tenantId = Some(TENANT_1))
+        }
+
+        Get("/v0/records/testId?aspect=test") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.NotFound
         }
       }
 
       it("can modify an aspect") { param =>
         val aspectDefinition = AspectDefinition("test", "test", None)
-        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val record = Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("bar"))))
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val patch = JsonPatch(Replace(Pointer.root / "aspects" / "test" / "foo", JsString("baz")))
-        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("baz"))), None, Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("baz"))), tenantId = Some(TENANT_1))
         }
 
-        Get("/v0/records/testId?aspect=test") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.BadRequest
+        }
+
+        Get("/v0/records/testId?aspect=test") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("baz"))), None, Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("baz"))), tenantId = Some(TENANT_1))
+        }
+
+        Get("/v0/records/testId?aspect=test") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.NotFound
         }
       }
 
       it("can add a new property to an aspect") { param =>
         val aspectDefinition = AspectDefinition("test", "test", None)
-        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val record = Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("bar"))))
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val patch = JsonPatch(Add(Pointer.root / "aspects" / "test" / "newprop", JsString("test")))
-        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("bar"), "newprop" -> JsString("test"))), None, Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("bar"), "newprop" -> JsString("test"))), tenantId = Some(TENANT_1))
         }
 
-        Get("/v0/records/testId?aspect=test") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.BadRequest
+        }
+
+        Get("/v0/records/testId?aspect=test") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("bar"), "newprop" -> JsString("test"))), None, Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("bar"), "newprop" -> JsString("test"))), tenantId = Some(TENANT_1))
+        }
+
+        Get("/v0/records/testId?aspect=test") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.NotFound
         }
       }
 
       it("can remove an aspect") { param =>
         val aspectDefinition = AspectDefinition("test", "test", None)
-        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val record = Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("bar"))))
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val patch = JsonPatch(Remove(Pointer.root / "aspects" / "test"))
-        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual Record("testId", "testName", Map(), None, Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual Record("testId", "testName", Map(), tenantId = Some(TENANT_1))
         }
 
-        Get("/v0/records/testId?optionalAspect=test") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.BadRequest
+        }
+
+        Get("/v0/records/testId?optionalAspect=test") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual Record("testId", "testName", Map(), None, Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual Record("testId", "testName", Map(), tenantId = Some(TENANT_1))
+        }
+
+        Get("/v0/records/testId?optionalAspect=test") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.NotFound
         }
       }
 
       it("can remove a property from an aspect") { param =>
         val aspectDefinition = AspectDefinition("test", "test", None)
-        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val record = Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("bar"), "newprop" -> JsString("test"))))
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val patch = JsonPatch(Remove(Pointer.root / "aspects" / "test" / "newprop"))
-        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("bar"))), None, Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("bar"))), tenantId = Some(TENANT_1))
         }
 
-        Get("/v0/records/testId?optionalAspect=test") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.BadRequest
+        }
+
+        Get("/v0/records/testId?optionalAspect=test") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("bar"))), None, Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("bar"))), tenantId = Some(TENANT_1))
+        }
+
+        Get("/v0/records/testId?optionalAspect=test") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.NotFound
         }
       }
 
       it("supports Move within an aspect") { param =>
         val aspectDefinition = AspectDefinition("test", "test", None)
-        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val record = Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("bar"))))
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val patch = JsonPatch(Move(Pointer.root / "aspects" / "test" / "foo", Pointer.root / "aspects" / "test" / "bar"))
-        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject("bar" -> JsString("bar"))), None, Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject("bar" -> JsString("bar"))), tenantId = Some(TENANT_1))
         }
 
-        Get("/v0/records/testId?optionalAspect=test") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.BadRequest
+        }
+
+        Get("/v0/records/testId?optionalAspect=test") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject("bar" -> JsString("bar"))), None, Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject("bar" -> JsString("bar"))), tenantId = Some(TENANT_1))
+        }
+
+        Get("/v0/records/testId?optionalAspect=test") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.NotFound
         }
       }
 
       it("supports Copy within an aspect") { param =>
         val aspectDefinition = AspectDefinition("test", "test", None)
-        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val record = Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("bar"))))
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val patch = JsonPatch(Copy(Pointer.root / "aspects" / "test" / "foo", Pointer.root / "aspects" / "test" / "bar"))
-        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("bar"), "bar" -> JsString("bar"))), None, Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("bar"), "bar" -> JsString("bar"))), tenantId = Some(TENANT_1))
         }
 
-        Get("/v0/records/testId?optionalAspect=test") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.BadRequest
+        }
+
+        Get("/v0/records/testId?optionalAspect=test") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("bar"), "bar" -> JsString("bar"))), None, Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual Record("testId", "testName", Map("test" -> JsObject("foo" -> JsString("bar"), "bar" -> JsString("bar"))), tenantId = Some(TENANT_1))
+        }
+
+        Get("/v0/records/testId?optionalAspect=test") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.NotFound
         }
       }
 
       it("evaluates Test operations") { param =>
         val A = AspectDefinition("A", "A", None)
-        param.asAdmin(Post("/v0/aspects", A)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/aspects", A)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val record = Record("testId", "testName", Map("A" -> JsObject("foo" -> JsString("bar"))))
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val patchSuccess = JsonPatch(Test(Pointer.root / "aspects" / "A" / "foo", JsString("bar")))
-        param.asAdmin(Patch("/v0/records/testId", patchSuccess)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patchSuccess)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
-          responseAs[Record] shouldEqual Record("testId", "testName", Map("A" -> JsObject("foo" -> JsString("bar"))), None, Some(SINGLE_TENANT_ID))
+          responseAs[Record] shouldEqual Record("testId", "testName", Map("A" -> JsObject("foo" -> JsString("bar"))), tenantId = Some(TENANT_1))
+        }
+
+        param.asAdmin(Patch("/v0/records/testId", patchSuccess)) ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+          status shouldEqual StatusCodes.BadRequest
         }
 
         val patchFail = JsonPatch(Test(Pointer.root / "aspects" / "A" / "foo", JsString("not this value")))
-        param.asAdmin(Patch("/v0/records/testId", patchFail)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patchFail)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.BadRequest
           responseAs[BadRequest].message should include("test failed")
         }
@@ -1499,22 +1766,22 @@ class RecordsServiceSpec extends ApiSpec {
 
       it("does not support Move between aspects") { param =>
         val A = AspectDefinition("A", "A", None)
-        param.asAdmin(Post("/v0/aspects", A)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/aspects", A)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val B = AspectDefinition("B", "B", None)
-        param.asAdmin(Post("/v0/aspects", B)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/aspects", B)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val record = Record("testId", "testName", Map("A" -> JsObject("foo" -> JsString("bar")), "B" -> JsObject()))
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val patch = JsonPatch(Move(Pointer.root / "aspects" / "A" / "foo", Pointer.root / "aspects" / "B" / "foo"))
-        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.BadRequest
           responseAs[BadRequest].message should include("two different aspects")
         }
@@ -1522,22 +1789,22 @@ class RecordsServiceSpec extends ApiSpec {
 
       it("does not support Copy between aspects") { param =>
         val A = AspectDefinition("A", "A", None)
-        param.asAdmin(Post("/v0/aspects", A)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/aspects", A)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val B = AspectDefinition("B", "B", None)
-        param.asAdmin(Post("/v0/aspects", B)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/aspects", B)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val record = Record("testId", "testName", Map("A" -> JsObject("foo" -> JsString("bar")), "B" -> JsObject()))
-        param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.OK
         }
 
         val patch = JsonPatch(Copy(Pointer.root / "aspects" / "A" / "foo", Pointer.root / "aspects" / "B" / "foo"))
-        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+        param.asAdmin(Patch("/v0/records/testId", patch)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
           status shouldEqual StatusCodes.BadRequest
           responseAs[BadRequest].message should include("two different aspects")
         }
@@ -1553,18 +1820,23 @@ class RecordsServiceSpec extends ApiSpec {
       describe("by id") {
         it("can delete a record without any aspects") { param =>
           val record = Record("without", "without", Map())
-          param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          param.asAdmin(Delete("/v0/records/without")) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          param.asAdmin(Delete("/v0/records/without")) ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            responseAs[DeleteResult].deleted shouldBe false
+          }
+
+          param.asAdmin(Delete("/v0/records/without")) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             responseAs[DeleteResult].deleted shouldBe true
           }
         }
 
         it("returns 200 and deleted=false when asked to delete a record that doesn't exist") { param =>
-          param.asAdmin(Delete("/v0/records/doesnotexist")) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          param.asAdmin(Delete("/v0/records/doesnotexist")) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             responseAs[DeleteResult].deleted shouldBe false
           }
@@ -1572,16 +1844,21 @@ class RecordsServiceSpec extends ApiSpec {
 
         it("can delete a record with an aspect") { param =>
           val aspectDefinition = AspectDefinition("test", "test", None)
-          param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
           val record = Record("with", "with", Map("test" -> JsObject()))
-          param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
-          param.asAdmin(Delete("/v0/records/with")) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          param.asAdmin(Delete("/v0/records/with")) ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            responseAs[DeleteResult].deleted shouldBe false
+          }
+
+          param.asAdmin(Delete("/v0/records/with")) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             responseAs[DeleteResult].deleted shouldBe true
           }
@@ -1607,7 +1884,7 @@ class RecordsServiceSpec extends ApiSpec {
           val lines = try source.mkString finally source.close()
 
           val aspectDefinition = AspectDefinition("source", "source", Some(lines.parseJson.asJsObject))
-          param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
@@ -1627,22 +1904,33 @@ class RecordsServiceSpec extends ApiSpec {
           val all = List(noTagNoSource, noTagWrongSource, noTagRightSource, wrongTagNoSource, wrongTagWrongSource,
             wrongTagRightSource, rightTagNoSource, rightTagWrongSource, rightTagRightSource1)
 
-          all.foreach(record => param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          all.foreach(record => param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
           })
 
-          Get("/v0/records") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val res = responseAs[RecordsPage[Record]]
             res.records.length shouldEqual all.length
           }
 
-          param.asAdmin(Delete("/v0/records?sourceTagToPreserve=righttag&sourceId=right")) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records") ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            val res = responseAs[RecordsPage[Record]]
+            res.records.length shouldEqual 0
+          }
+
+          param.asAdmin(Delete("/v0/records?sourceTagToPreserve=righttag&sourceId=right")) ~> addTenantIdHeader(TENANT_2) ~> param.api(role).routes ~> check {
+            status shouldEqual StatusCodes.OK
+            responseAs[MultipleDeleteResult].count shouldEqual 0
+          }
+
+          param.asAdmin(Delete("/v0/records?sourceTagToPreserve=righttag&sourceId=right")) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             responseAs[MultipleDeleteResult].count shouldEqual 2
           }
 
-          Get("/v0/records") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val res = responseAs[RecordsPage[Record]]
             res.records.length shouldEqual (all.length - 2)
@@ -1651,15 +1939,16 @@ class RecordsServiceSpec extends ApiSpec {
           }
 
           List("wrongtag-rightsource", "notag-rightsource").foreach { recordId =>
-            Get(s"/v0/records/$recordId") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+            Get(s"/v0/records/$recordId") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
               status shouldEqual StatusCodes.NotFound
             }
 
-            Get(s"/v0/records/$recordId/history") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+            Get(s"/v0/records/$recordId/history") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
               status shouldEqual StatusCodes.OK
               val res = responseAs[EventsPage]
 
               val deleteRecordEvents = res.events.filter(event => event.eventType == EventType.DeleteRecord)
+
               deleteRecordEvents.length shouldEqual 1
               val deletedRecordEvent = deleteRecordEvents.head.data.convertTo[Map[String, JsValue]]
               deletedRecordEvent("recordId") shouldEqual JsString(recordId)
@@ -1682,7 +1971,7 @@ class RecordsServiceSpec extends ApiSpec {
             Success(1)
           }
 
-          param.asAdmin(Delete("?sourceTagToPreserve=righttag&sourceId=right")) ~> addSingleTenantIdHeader ~> mockedApi.route ~> check {
+          param.asAdmin(Delete("?sourceTagToPreserve=righttag&sourceId=right")) ~> addTenantIdHeader(TENANT_1) ~> mockedApi.route ~> check {
             status shouldEqual StatusCodes.Accepted
           }
         }
@@ -1697,7 +1986,7 @@ class RecordsServiceSpec extends ApiSpec {
           val lines = try source.mkString finally source.close()
 
           val aspectDefinition = AspectDefinition("source", "source", Some(lines.parseJson.asJsObject))
-          param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          param.asAdmin(Post("/v0/aspects", aspectDefinition)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
           }
 
@@ -1705,22 +1994,22 @@ class RecordsServiceSpec extends ApiSpec {
 
           val all = List(rightTagRightSource1)
 
-          all.foreach(record => param.asAdmin(Post("/v0/records", record)) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          all.foreach(record => param.asAdmin(Post("/v0/records", record)) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
           })
 
-          Get("/v0/records") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val res = responseAs[RecordsPage[Record]]
             res.records.length shouldEqual all.length
           }
 
-          param.asAdmin(Delete("/v0/records?sourceTagToPreserve=righttag&sourceId=right")) ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          param.asAdmin(Delete("/v0/records?sourceTagToPreserve=righttag&sourceId=right")) ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             responseAs[MultipleDeleteResult].count shouldEqual 0
           }
 
-          Get("/v0/records") ~> addSingleTenantIdHeader ~> param.api(role).routes ~> check {
+          Get("/v0/records") ~> addTenantIdHeader(TENANT_1) ~> param.api(role).routes ~> check {
             status shouldEqual StatusCodes.OK
             val res = responseAs[RecordsPage[Record]]
             res.records.length shouldEqual 1
